Alert user before closing the createItemModal form
The form data inside the `createItemModal` was getting lost when user
hit the `cancel` button. This could lead to a bad UX.

This fix will identifies if form data has changed from its initial
values before closing the modal. If yes, then it alerts the user for
confirmation. Otherwise, will close the modal.

Closes: #2822